extract slash commands into commands/ and AniList client into lib/
Each command is now its own file under commands/ with a {data, ephemeral, execute} shape; commands/index.js auto-loads them. /help auto-renders from the registry. Net: index.js drops ~500 lines, deploy-commands.js drops ~50.
